requirements.txt
streamlit==1.32.0
yfinance==0.2.37
pandas==2.2.1
plotly==5.20.0
requests==2.31.0
streamlit-autorefresh==1.0.1
feedparser==6.0.11
google-generativeai==0.8.3
